feat: Separate activity filters into individual components
