+++ src/activity-feed/__tests__/ActivityFeedFilters.test.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { shallow } from 'enzyme'
 
 import ActivityFeedFilters from '../ActivityFeedFilters'
 
@@ -10,50 +9,13 @@ describe('ActivityFeedFilters', () => {
       const tree = renderer
         .create(
           <ActivityFeedFilters
+            isActivityTypeFilterEnabled={false}
             onShowDetailsClick={() => {}}
             showDetails={false}
           />
         )
         .toJSON()
       expect(tree).toMatchSnapshot()
-    })
-  })
-
-  describe('when the details for all activities are shown', () => {
-    test('renders filters', () => {
-      const tree = renderer
-        .create(
-          <ActivityFeedFilters
-            onShowDetailsClick={() => {}}
-            showDetails={true}
-          />
-        )
-        .toJSON()
-      expect(tree).toMatchSnapshot()
-    })
-  })
-
-  describe('when the "Show details for all activities" checkbox is clicked', () => {
-    const onShowDetailsClickMock = jest.fn()
-
-    beforeEach(() => {
-      const wrapper = shallow(
-        <ActivityFeedFilters
-          onShowDetailsClick={onShowDetailsClickMock}
-          showDetails={false}
-        />
-      )
-      wrapper.find('Checkbox').simulate('change', {
-        target: { checked: true },
-      })
-    })
-
-    afterEach(() => {
-      onShowDetailsClickMock.mockReset()
-    })
-
-    test('should call onShowDetailsClick', () => {
-      expect(onShowDetailsClickMock.mock.calls.length).toEqual(1)
     })
   })
 })
